feat: move main files to src directory
