--- __init__.py
@@ -1,3 +1,0 @@
-from World import *
-from DataStructures import *
-from GlobalPlanning import *
