+++ experiments/experiment_large.py
@@ -1,10 +1,12 @@
 
+# external imports
 import os
-from Experiment import *
-from unittests import *
 from experiments.large.config import *
-
 from matplotlib.ticker import MaxNLocator
+
+# internal imports
+from src.Experiment import *
+
 
 ##################################
 ## NO NEED TO MAKE CHANGES HERE ##
@@ -24,11 +26,11 @@ exp_res_file = os.path.join(exp_dir, exp_res_filename + pickle_extension)
 
 # adapt the optimization parameters
 op = OptimizationParameters()
-op._kkt_tolerance = 1e-1
-op._sim_to_steady_state_tol = 1e-1
-op._optimization_iters = 100
-op._steady_state_iters = 1
-op._beta = 0.95
+op.kkt_tolerance = 1e-1
+op.sim_to_steady_state_tol = 1e-1
+op.optimization_iters = 100
+op.steady_state_iters = 1
+op.beta = 0.95
 
 if not os.path.exists(exp_dir):
     os.makedirs(exp_dir)
